Some modularization
