+++ app/src/components/StripeCard.tsx
@@ -1,28 +1,7 @@
 import React from "react";
-import styled from "styled-components";
+import { Wrapper } from "./";
+import { Wallet } from "../types/stripe";
 
-export interface Wallet {
-  billing_details: {
-    name: string;
-  };
-  card: {
-    brand: string;
-    country: string;
-    exp_month: number;
-    exp_year: number;
-    last4: string;
-  };
-  id: string;
-}
-const Wrapper = styled.div`
-  width: 100%;
-  height: 100px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
-  border: 1px solid black;
-`;
 const StripeCard = ({
   paymentMethod,
   handleSelect,
