+++ app/src/components/StripeItem.tsx
@@ -1,43 +1,13 @@
 import React from "react";
-import styled from "styled-components";
-const Wrapper = styled.div`
-  width: 100%;
-  height: 100px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
-  border: 1px solid black;
-`;
-
-export interface Product {
-  id: string;
-  name: string;
-  prices: Price[];
-}
-
-export interface Price {
-  id: string;
-  product: string;
-  currency: string;
-  unit_amount: number;
-  recurring: {
-    interval: "month" | "year";
-    interval_count: number;
-  };
-}
+import { Wrapper } from "./";
+import { Price } from "../types/stripe";
 
 const StripePrice = ({
   price,
   handleSubscribe,
   currentPrice,
 }: {
-  price: {
-    id: string;
-    currency: string;
-    unit_amount: number;
-    recurring: { interval_count: number; interval: string };
-  };
+  price: Price;
   handleSubscribe: () => void;
   currentPrice?: string;
 }) => {
