+++ app/src/components/Wrapper.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+const Wrapper = styled.div`
+  width: 100%;
+  height: 100px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  align-items: center;
+  border: 1px solid black;
+`;
+
+export default Wrapper;
