+++ app/src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as Navbar } from "./Navbar";
+export { default as Wrapper } from "./Wrapper";
